Don't use long to store pointers
long may be to short to store pointers on 64 bit systems and therefore cause
build failures on Windows.

+++ src/Assembler/assembler.h
@@ -114,7 +114,7 @@ Done:
     }
 #else
 
-    long beginOffset = (long)src & 3;
+    long long beginOffset = (long long)src & 3;
     char *readPtr = (char*)src - beginOffset;
     char *writePtr = (char*)dest;
